refactor: undeprecate `@remirror/dev`
