+++ .changeset/purple-moons-cross.md
@@ -1,10 +1,5 @@
 ---
-'@remirror/cli': major
 '@remirror/core-extensions': major
-'@remirror/dev': major
-'@remirror/editor-markdown': major
-'@remirror/editor-social': major
-'@remirror/editor-wysiwyg': major
 '@remirror/react-hooks': major
 '@remirror/react-node-view': major
 '@remirror/react-portals': major
