--- @remirror/react-dev/src/index.ts
@@ -1,1 +1,0 @@
-export * from './react-dev';
