--- deprecated/dev/CHANGELOG.md
@@ -1,26 +1,0 @@
-# @remirror/dev
-
-## 0.11.0
-
-### Patch Changes
-
-- Updated dependencies [026d4238]
-- Updated dependencies [69d00c62]
-  - @remirror/react@0.11.0
-
-## 0.7.4
-
-### Patch Changes
-
-- 7380e18f: Update repository url from ifiokjr/remirror to remirror/remirror to reflect new GitHub
-  organisation.
-- Updated dependencies [7380e18f]
-  - @remirror/react@0.7.5
-
-## 0.7.3
-
-### Patch Changes
-
-- 5f85c0de: Bump a new version to test out the changeset API.
-- Updated dependencies [5f85c0de]
-  - @remirror/react@0.7.3
