--- deprecated/dev/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "@remirror/dev",
-  "version": "0.13.2",
-  "description": "Development utilities for remirror",
-  "homepage": "https://github.com/remirror/remirror/tree/next/@remirror/react-dev",
-  "repository": "https://github.com/remirror/remirror/tree/master/deprecated/dev",
-  "license": "MIT",
-  "contributors": [
-    "Ifiok Jr. <ifiokotung@gmail.com>"
-  ],
-  "publishConfig": {
-    "access": "public"
-  }
-}
