--- deprecated/dev/readme.md
@@ -1,16 +1,0 @@
-# @remirror/dev
-
-This package is **deprecated** and has been moved to `@remirror/react-dev` package.
-
-It is now accessible with the following installation.
-
-```bash
-# Yarn
-yarn add @remirror/react-dev
-
-# PNPM
-pnpm add @remirror/react-dev
-
-# npm
-npm install @remirror/react-dev
-```
